@@ -1,7 +1,9 @@
 <?php
 
+declare(strict_types=1);
+
+use Illuminate\Contracts\View\Factory;
+use Illuminate\Contracts\View\View;
 use Illuminate\Support\Facades\Route;
 
-Route::get('/', function () {
-    return view('welcome');
-});
+Route::get('/', fn (): View|Factory => view('welcome'));
