@@ -1,16 +1,19 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Tests\Unit;
 
 use PHPUnit\Framework\TestCase;
 
-class ExampleTest extends TestCase
+final class ExampleTest extends TestCase
 {
     /**
      * A basic test example.
      */
     public function test_that_true_is_true(): void
     {
+        /** @phpstan-ignore-next-line */
         $this->assertTrue(true);
     }
 }
